move torrent creation
