+++ src/trackers/HDB.py
@@ -12,6 +12,7 @@ from src.exceptions import *  # noqa F403
 from src.console import console
 from datetime import datetime
 from torf import Torrent
+from src.torrentcreate import CustomTorrent, torf_cb
 
 
 class HDB():
@@ -210,9 +211,8 @@ class HDB():
             if each == "EXIT":
                 console.print("[bold red]Something didn't map correctly, or this content is not allowed on HDB")
                 return
-        if "Dual-Audio" in meta['audio']:
-            if not (meta['anime'] or meta['is_disc']):
-                console.print("[bold red]Dual-Audio Encodes are not allowed for non-anime and non-disc content")
+        if "Dual-Audio" in meta['audio'] and meta['is_disc'] not in ("BDMV", "HDDVD", "DVD"):
+            console.print("[bold red]Dual-Audio Encodes are not allowed")
             return
 
         # Download new .torrent from site
@@ -224,10 +224,6 @@ class HDB():
         if torrent.piece_size > 16777216:  # 16 MiB in bytes
             console.print("[red]Piece size is OVER 16M and does not work on HDB. Generating a new .torrent")
 
-            # Import Prep and regenerate the torrent with 16 MiB piece size limit
-            from src.prep import Prep
-            prep = Prep(screens=meta['screens'], img_host=meta['imghost'], config=self.config)
-
             if meta['is_disc'] == 1:
                 include = []
                 exclude = []
@@ -236,7 +232,7 @@ class HDB():
                 exclude = ["*.*", "*sample.mkv", "!sample*.*"]
 
             # Create a new torrent with piece size explicitly set to 16 MiB
-            new_torrent = prep.CustomTorrent(
+            new_torrent = CustomTorrent(
                 meta=meta,
                 path=Path(meta['path']),
                 trackers=["https://fake.tracker"],
@@ -255,7 +251,7 @@ class HDB():
 
             # Validate and write the new torrent
             new_torrent.validate_piece_size()
-            new_torrent.generate(callback=prep.torf_cb, interval=5)
+            new_torrent.generate(callback=torf_cb, interval=5)
             new_torrent.write(torrent_path, overwrite=True)
 
         # Proceed with the upload process
